+++ src/Traits/ComponentUtilities.php
@@ -31,6 +31,17 @@ trait ComponentUtilities
     protected $tdAttributesCallback;
     protected $collapsingColumnsStatus = true;
     protected string $emptyMessage = 'No items found. Try to broaden your search.';
+    protected array $additionalSelects = [];
+    protected array $configurableAreas = [
+        'toolbar-left-start' => null,
+        'toolbar-left-end' => null,
+        'toolbar-right-start' => null,
+        'toolbar-right-end' => null,
+        'before-toolbar' => null,
+        'after-toolbar' => null,
+        'before-pagination' => null,
+        'after-pagination' => null,
+    ];
 
     /**
      * Set the custom query string array for this specific table