+++ src/Traits/Helpers/ComponentHelpers.php
@@ -303,4 +303,40 @@ trait ComponentHelpers
     {
         return $this->trUrlTargetCallback ? call_user_func($this->trUrlTargetCallback, $row) : null;
     }
+
+    /**
+     * @return array
+     */
+    public function getAdditionalSelects(): array
+    {
+        return $this->additionalSelects;
+    }
+
+    /**
+     * @return array
+     */
+    public function getConfigurableAreas(): array
+    {
+        return $this->configurableAreas;
+    }
+
+    /**
+     * @param  string  $area
+     *
+     * @return bool
+     */
+    public function hasConfigurableAreaFor(string $area): bool
+    {
+        return isset($this->configurableAreas[$area]) && $this->getConfigurableAreaFor($area) !== null;
+    }
+
+    /**
+     * @param  string  $area
+     *
+     * @return string|null
+     */
+    public function getConfigurableAreaFor(string $area): ?string
+    {
+        return $this->configurableAreas[$area] ?? null;
+    }
 }