+++ tests/Traits/Helpers/ComponentHelpersTest.php
@@ -152,4 +152,35 @@ class ComponentHelpersTest extends TestCase
 
         $this->assertTrue($this->basicTable->hasTableRowUrl());
     }
+
+    /** @test */
+    public function can_get_additional_selects(): void
+    {
+        $this->assertEquals([], $this->basicTable->getAdditionalSelects());
+
+        $this->basicTable->setAdditionalSelects(['id', 'name']);
+
+        $this->assertEquals(['id', 'name'], $this->basicTable->getAdditionalSelects());
+    }
+
+    /** @test */
+    public function can_get_configurable_areas(): void
+    {
+        $this->assertEquals([
+            'toolbar-left-start' => null,
+            'toolbar-left-end' => null,
+            'toolbar-right-start' => null,
+            'toolbar-right-end' => null,
+            'before-toolbar' => null,
+            'after-toolbar' => null,
+            'before-pagination' => null,
+            'after-pagination' => null,
+        ], $this->basicTable->getConfigurableAreas());
+
+        $this->basicTable->setConfigurableAreas([
+            'toolbar-left-start' => 'includes.areas.toolbar-left-start',
+        ]);
+
+        $this->assertEquals('includes.areas.toolbar-left-start', $this->basicTable->getConfigurableAreaFor('toolbar-left-start'));
+    }
 }